Fix reading of bad relocation type (#1672)
Fixes: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=30792

+++ src/common.h
@@ -420,7 +420,7 @@ void InitStdio();
 extern const char* g_kind_name[];
 
 static WABT_INLINE const char* GetKindName(ExternalKind kind) {
-  return static_cast<int>(kind) < kExternalKindCount
+  return static_cast<size_t>(kind) < kExternalKindCount
     ? g_kind_name[static_cast<size_t>(kind)]
     : "<error_kind>";
 }
@@ -430,7 +430,7 @@ static WABT_INLINE const char* GetKindName(ExternalKind kind) {
 extern const char* g_reloc_type_name[];
 
 static WABT_INLINE const char* GetRelocTypeName(RelocType reloc) {
-  return static_cast<int>(reloc) < kRelocTypeCount
+  return static_cast<size_t>(reloc) < kRelocTypeCount
     ? g_reloc_type_name[static_cast<size_t>(reloc)]
     : "<error_reloc_type>";
 }
